requirements.txt
fastapi
uvicorn
pymongo
python-jose
passlib
bcrypt
python-multipart